requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
alembic==1.13.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
    # via -r requirements.in
bcrypt==4.2.0
    # via -r requirements.in
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
fastapi==0.112.0
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.7
    # via anyio
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via mako
passlib==1.7.4
    # via -r requirements.in
pydantic==2.8.2
    # via fastapi
pydantic-core==2.20.1
    # via pydantic
pyjwt==2.9.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via uvicorn
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via anyio
sqlalchemy[asyncio]==2.0.32
    # via
    #   -r requirements.in
    #   alembic
starlette==0.37.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn[standard]==0.30.6
    # via -r requirements.in
watchfiles==0.23.0
    # via uvicorn
websockets==12.0
    # via
    #   -r requirements.in
    #   uvicorn
